final_deployment/requirements.txt
flask
pyserial
numpy
sounddevice
soundfile
